[feat] set up flask backend and api routes to trigger webscraper database actions
